// frontend/src/shared/lib/api/model/structureSearchResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Chem-Vault
 * OpenAPI spec version: 0.1.0
 */
import type { StructureSearchResponseMolecules } from './structureSearchResponseMolecules';
import type { StructureSearchResponseSimilarityResults } from './structureSearchResponseSimilarityResults';

/**
 * Wrapper for structure search results (exact, substructure, or similarity).
 */
export interface StructureSearchResponse {
  search_type: string;
  molecules?: StructureSearchResponseMolecules;
  similarity_results?: StructureSearchResponseSimilarityResults;
  count: number;
}