// frontend/src/shared/lib/api/model/addIdentifierResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Cellar
 * OpenAPI spec version: 0.1.0
 */
import type { IdentifierResponse } from './identifierResponse';
import type { MirrorSummaryResponse } from './mirrorSummaryResponse';

export interface AddIdentifierResponse {
  identifiers: IdentifierResponse[];
  mirror_summary: MirrorSummaryResponse;
}